Enhanced debugging

+++ jenkins/setup.sh
@@ -15,20 +15,28 @@ init() {
         esac
     fi
 
-    rm -fr $SHARED_DIRECTORY
-    
-    # Create the shared directory as well as the Maven .m2/repository directory and the workspace
-    for d in $SHARED_DIRECTORY $SHARED_DIRECTORY/.m2/repository $SHARED_DIRECTORY/agent/workspace
-    do
-        test -d $d || mkdir -p $d
-        chmod -R 755 $d
-    done
-
-    if [ ! -d $SHARED_DIRECTORY/.ssh ]
+    if [ $NFS -ne 0 ]
     then
-        mkdir -m 700 $SHARED_DIRECTORY/.ssh
-        ssh-keyscan github.com > $SHARED_DIRECTORY/.ssh/known_hosts
-        chmod 700 $SHARED_DIRECTORY/.ssh/known_hosts
+        if ! printenv SHARED_DIRECTORY
+        then
+            export SHARED_DIRECTORY=~/nfs/jenkins/home
+        fi
+    
+        rm -fr $SHARED_DIRECTORY
+    
+        # Create the shared directory as well as the Maven .m2/repository directory and the workspace
+        for d in $SHARED_DIRECTORY $SHARED_DIRECTORY/.m2/repository $SHARED_DIRECTORY/agent/workspace
+        do
+            test -d $d || mkdir -p $d
+            chmod -R 755 $d
+        done
+
+        if [ ! -d $SHARED_DIRECTORY/.ssh ]
+        then
+            mkdir -m 700 $SHARED_DIRECTORY/.ssh
+            ssh-keyscan github.com > $SHARED_DIRECTORY/.ssh/known_hosts
+            chmod 700 $SHARED_DIRECTORY/.ssh/known_hosts
+        fi
     fi
 
     # The docker-compose.yml is the common file.
